Add Save/Load feature to Oware
Add functionality to loadGame and saveGame methods which are called by
their respective commands outlined in the the project brief.

Add copy constructor for GameImpl.

Change GameImpl toString so it can easily be saved as a file. On review
efficiency may dictate that this is the wrong way to implement this as
moves are made often while saves less so.

Change captureAllSeeds() in GameImpl so that the generic type Board can
be used.

Change playGame() in GameManager to process GameImpl's new toString.

Complete LOAD and SAVE commands in manage(input, output) in
GameManagerImpl.

Create private checkValidIOCommand(String) to ensure LOAD and SAVE
commands are followed by exactly one other String (the file name).

Create private processBoardString(String) to take the toString of Board
and convert it to a BoardImpl object. (May make more sense to reflect
in the Javadoc that it expects a BoardImpl toString.